componentizando navbar
