+++ src/components/menuItems/menuItens.js
@@ -3,34 +3,35 @@ import * as Icons from 'react-icons/fa';
 export const navItems = [
   {
     id: 1,
+    title: 'Menu',
+    path: '/orders',
+    nName: 'nav-item',
+    sName: 'sidebar-item',
+    icon: <Icons.FaMoneyCheck />,
+  },
+  {
+    id: 2,
     title: 'Pedidos Prontos',
     path: '/waiter',
     nName: 'nav-item',
     sName: 'sidebar-item',
     icon: <Icons.FaFireAlt />,
   },
-  /* {
-    id: 2,
-    title: 'LogOut',
-    path: './',
-    nName: 'nav-item',
-    sName: 'sidebar-item',
-    icon: <Icons.FaDoorOpen />,
-  },
- {
-    id: 3,
-    title: 'Products',
-    path: './products',
-    nName: 'nav-item',
-    sName: 'sidebar-item',
-    icon: <Icons.FaCartArrowDown />,
-  },
-  {
-    id: 4,
-    title: 'Contact Us',
-    path: './contactus',
-    nName: 'nav-item',
-    sName: 'sidebar-item',
-    icon: <Icons.FaPhone />,
-  },*/
+
+//  {
+//     id: 3,
+//     title: 'Products',
+//     path: './products',
+//     nName: 'nav-item',
+//     sName: 'sidebar-item',
+//     icon: <Icons.FaCartArrowDown />,
+//   },
+//   {
+//     id: 4,
+//     title: 'Contact Us',
+//     path: './contactus',
+//     nName: 'nav-item',
+//     sName: 'sidebar-item',
+//     icon: <Icons.FaPhone />,
+//   },
 ];
