+++ src/pages/Kitchen/kitchen.css
@@ -6,7 +6,7 @@
 }
 
 .contentOrdersKitchen {
-  width: 28vw;
+  width: 31vw;
 }
 
 .contentOrdersKitchen h1 {
